llm_integrations/api/pyproject.toml
[tool.poetry]
name = "llm-api"
version = "0.1.0"
description = "PDF processing API using FastAPI and Google Gemini"
authors = ["Developer <developer@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.110.0"
uvicorn = "^0.28.0"
pydantic = "^2.7.0"
pymupdf = "^1.24.0"
python-multipart = "0.0.6"
python-dotenv = "^1.0.0"
requests = "^2.32.3"
google-genai = "1.13.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
